requirements.txt
absl-py==1.4.0
altgraph==0.17.3
asgiref==3.7.2
astroid==2.15.3
astunparse==1.6.3
async-timeout==4.0.2
autopep8==2.0.2
beautifulsoup4==4.12.2
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
comtypes==1.2.0
configobj==5.0.8
configparser==6.0.0
contourpy==1.0.7
cycler==0.11.0
decompyle3==3.9.1.dev0
dill==0.3.6
Django==4.2.4
docopt==0.6.2
et-xmlfile==1.1.0
etelemetry==0.3.0
filelock==3.12.2
fitz==0.0.1.dev2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Session==0.4.0
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
gast==0.4.0
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
h5py==3.8.0
httplib2==0.22.0
idna==3.4
iqoptionapi==6.8.9.1
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
libclang==16.0.0
looseversion==1.3.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
metatrader==0.0.1
MetaTrader5==5.0.45
ml-dtypes==0.1.0
mplfinance==0.12.9b7
neo4j==5.7.0
neo4j-driver==5.8.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.0
pandas-datareader==0.10.0
pandas-ta==0.3.14b0
pathlib==1.0.1
pefile==2023.2.7
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.2.0
proto-plus==1.22.2
protobuf==4.22.3
prov==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycryptodome==3.17
pydot==1.4.2
pydumpck==1.15.19
Pygments==2.15.1
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pylint==3.0.0a6
PyMuPDF==1.23.1
PyMuPDFb==1.23.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
pywifi==1.1.12
pywin32-ctypes==0.2.0
pyxnat==1.6
rdflib==7.0.0
redis==4.5.4
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scapy==2.5.0
scikit-learn==1.2.2
scipy==1.10.1
sgtpyutils==1.15.17
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
spark-parser==1.8.9
sqlparse==0.4.4
TA-Lib @ file:///C:/Users/Sam/Downloads/TA_Lib-0.4.24-cp310-cp310-win_amd64.whl
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tinyaes==1.0.4
tomli==2.0.1
tomlkit==0.11.7
tqdm==4.66.1
tradingview-ta==3.3.0
traits==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
uncompyle6==3.9.0
urllib3==1.26.15
websocket-client==0.56.0
Werkzeug==2.2.3
wrapt==1.14.1
xdis==6.0.5
yarg==0.1.9
